Split selection strategies code into files in its own directory
