+++ src/index.ts
@@ -9,8 +9,8 @@ export { DynamicClusterPool } from './pools/cluster/dynamic'
 export { FixedClusterPool } from './pools/cluster/fixed'
 export type { ClusterPoolOptions } from './pools/cluster/fixed'
 export type { IPool } from './pools/pool'
-export { WorkerChoiceStrategies } from './pools/selection-strategies'
-export type { WorkerChoiceStrategy } from './pools/selection-strategies'
+export { WorkerChoiceStrategies } from './pools/selection-strategies/selection-strategies-types'
+export type { WorkerChoiceStrategy } from './pools/selection-strategies/selection-strategies-types'
 export { DynamicThreadPool } from './pools/thread/dynamic'
 export { FixedThreadPool } from './pools/thread/fixed'
 export type { ThreadWorkerWithMessageChannel } from './pools/thread/fixed'
