+++ src/pools/selection-strategies/round-robin-worker-choice-strategy.ts
@@ -1,0 +1,40 @@
+import type { IWorker } from '../abstract-pool'
+import type { IPoolInternal } from '../pool-internal'
+import type { IWorkerChoiceStrategy } from './selection-strategies-types'
+
+/**
+ * Selects the next worker in a round robin fashion.
+ *
+ * @template Worker Type of worker which manages the strategy.
+ * @template Data Type of data sent to the worker. This can only be serializable data.
+ * @template Response Type of response of execution. This can only be serializable data.
+ */
+export class RoundRobinWorkerChoiceStrategy<
+  Worker extends IWorker,
+  Data,
+  Response
+> implements IWorkerChoiceStrategy<Worker> {
+  /**
+   * Index for the next worker.
+   */
+  private nextWorkerIndex: number = 0
+
+  /**
+   * Constructs a worker choice strategy that selects in a round robin fashion.
+   *
+   * @param pool The pool instance.
+   */
+  public constructor (
+    private readonly pool: IPoolInternal<Worker, Data, Response>
+  ) {}
+
+  /** @inheritdoc */
+  public choose (): Worker {
+    const chosenWorker = this.pool.workers[this.nextWorkerIndex]
+    this.nextWorkerIndex =
+      this.pool.workers.length - 1 === this.nextWorkerIndex
+        ? 0
+        : this.nextWorkerIndex + 1
+    return chosenWorker
+  }
+}
