+++ src/pools/selection-strategies/selection-strategies-utils.ts
@@ -1,0 +1,44 @@
+import type { IWorker } from '../abstract-pool'
+import type { IPoolInternal } from '../pool-internal'
+import { FairShareWorkerChoiceStrategy } from './fair-share-worker-choice-strategy'
+import { LessRecentlyUsedWorkerChoiceStrategy } from './less-recently-used-worker-choice-strategy'
+import { RoundRobinWorkerChoiceStrategy } from './round-robin-worker-choice-strategy'
+import type {
+  IWorkerChoiceStrategy,
+  WorkerChoiceStrategy
+} from './selection-strategies-types'
+import { WorkerChoiceStrategies } from './selection-strategies-types'
+
+/**
+ * Worker selection strategies helpers class.
+ */
+export class SelectionStrategiesUtils {
+  /**
+   * Get the worker choice strategy instance.
+   *
+   * @param pool The pool instance.
+   * @param workerChoiceStrategy The worker choice strategy.
+   * @returns The worker choice strategy instance.
+   */
+  public static getWorkerChoiceStrategy<
+    Worker extends IWorker,
+    Data,
+    Response
+  > (
+    pool: IPoolInternal<Worker, Data, Response>,
+    workerChoiceStrategy: WorkerChoiceStrategy = WorkerChoiceStrategies.ROUND_ROBIN
+  ): IWorkerChoiceStrategy<Worker> {
+    switch (workerChoiceStrategy) {
+      case WorkerChoiceStrategies.ROUND_ROBIN:
+        return new RoundRobinWorkerChoiceStrategy(pool)
+      case WorkerChoiceStrategies.LESS_RECENTLY_USED:
+        return new LessRecentlyUsedWorkerChoiceStrategy(pool)
+      case WorkerChoiceStrategies.FAIR_SHARE:
+        return new FairShareWorkerChoiceStrategy(pool)
+      default:
+        throw new Error(
+          `Worker choice strategy '${workerChoiceStrategy}' not found`
+        )
+    }
+  }
+}
